Moved components content to app.py
